+++ app/_components/StartScreen.js
@@ -1,5 +1,5 @@
 import { makeStyles, tokens, shorthands } from '@fluentui/react-components';
-import { useTasks } from '../src/contexts/TasksContext';
+import { useTasks } from '../../src/contexts/TasksContext';
 
 const useStyles = makeStyles({
   startScreen: {
