+++ app/_components/User.js
@@ -16,7 +16,7 @@ import {
   MenuItem,
   useRestoreFocusTarget,
 } from '@fluentui/react-components';
-import { useTasks } from '../src/contexts/TasksContext';
+import { useTasks } from '../../src/contexts/TasksContext';
 import {
   CheckboxChecked20Regular,
   CheckboxUnchecked20Regular,
