+++ app/_components/UserSelector.js
@@ -6,7 +6,7 @@ import {
   makeStyles,
   tokens,
 } from '@fluentui/react-components';
-import { useTasks } from '../src/contexts/TasksContext';
+import { useTasks } from '../../src/contexts/TasksContext';
 import { PersonAddRegular } from '@fluentui/react-icons';
 import UserList from './UserList';
 import UpdateUserForm from './UpdateUserForm';
